Development Update 2
Event class:
-changed wait time to be of type float for greater accuracy
Simulation class:
-changed addObject function to return the resulting Object list iterator
-changed Time to be of type float for greater accuracy
-moved getSiteIndex function to protected status so that it is
accessible by derived simulations classes
-corrected logMSG function to take an ostringstream instead of
stringstream
Utils:
-improved seeding of random number generators
-added inpow function for floats

+++ Event.cpp
@@ -19,7 +19,7 @@ Coords Event::getDestCoords(){
     return coords_dest;
 }
 
-double Event::getWaitTime(){
+float Event::getWaitTime(){
     return wait_time;
 }
 
@@ -50,7 +50,7 @@ void Event::setDestCoords(const Coords& coords){
     coords_dest.z = coords.z;
 }
 
-void Event::setWaitTime(const double time){
+void Event::setWaitTime(const float time){
     wait_time = time;
 }
